+++ tests/cascadetimercall.cpp
@@ -55,7 +55,7 @@ int main(int argc, char* argv[]) {
     std::cout << "Timer started\n";
     std::mutex m;
     std::unique_lock lock(m);
-    cv.wait_for(lock, std::chrono::seconds(120));
+    cv.wait_for(lock, std::chrono::seconds(20));
     if (!called) {
         return EXIT_FAILURE;
     }